requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.44.0
django-filter==2.4.0
django-heroku==0.3.1
django-nose==1.4.7
django-sslserver==0.22
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
Markdown==3.3.4
nose==1.3.7
oauthlib==3.1.0
psycopg2-binary==2.8.6
psycopg2==2.8.6
pycparser==2.20
Pygments==2.8.1
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
uWSGI>=2.0.18,<2.1
pylint==2.8.2
